🚛 certif: reorganizing folders components for sessions
