+++ certif/app/components/sessions/session-details/controls-links.gjs
