+++ certif/app/components/sessions/session-details/details-modal-row.gjs
