+++ certif/app/components/sessions/session-details/header.gjs
