+++ certif/app/components/sessions/session-details/index.gjs
