+++ certif/app/components/sessions/session-details/nav.gjs
