+++ certif/app/templates/authenticated/sessions/details/certification-candidates.hbs
@@ -4,7 +4,7 @@
     <CertificationCandidatesSco @sessionId={{this.currentSession.id}} />
   {{/unless}}
   {{#if this.hasOneOrMoreCandidates}}
-    <EnrolledCandidates
+    <Sessions::SessionDetails::EnrolledCandidates
       @shouldDisplayPrescriptionScoStudentRegistrationFeature={{this.shouldDisplayPrescriptionScoStudentRegistrationFeature}}
       @sessionId={{this.currentSession.id}}
       @certificationCandidates={{this.certificationCandidates}}
@@ -21,7 +21,7 @@
     @importAllowed={{this.importAllowed}}
     @fetchCandidatesImportTemplateAction={{this.fetchCandidatesImportTemplate}}
   />
-  <EnrolledCandidates
+  <Sessions::SessionDetails::EnrolledCandidates
     @shouldDisplayPrescriptionScoStudentRegistrationFeature={{this.shouldDisplayPrescriptionScoStudentRegistrationFeature}}
     @sessionId={{this.currentSession.id}}
     @certificationCandidates={{this.certificationCandidates}}
