+++ certif/tests/integration/components/sessions/session-list-header_test.gjs
@@ -1,6 +1,6 @@
 import { render } from '@1024pix/ember-testing-library';
 import Service from '@ember/service';
-import PanelHeader from 'pix-certif/components/sessions/panel-header';
+import SessionListHeader from 'pix-certif/components/sessions/session-list-header';
 import { module, test } from 'qunit';
 
 import setupIntlRenderingTest from '../../../helpers/setup-intl-rendering';
@@ -23,7 +23,7 @@ module('Integration | Component | panel-header', function (hooks) {
     this.owner.register('service:current-user', CurrentUserStub);
 
     // when
-    const { getByRole } = await render(<template><PanelHeader /></template>);
+    const { getByRole } = await render(<template><SessionListHeader /></template>);
 
     // then
     assert.dom(getByRole('link', { name: 'Créer une session' })).exists();
@@ -45,7 +45,7 @@ module('Integration | Component | panel-header', function (hooks) {
       this.owner.register('service:current-user', CurrentUserStub);
 
       // when
-      const { queryByRole } = await render(<template><PanelHeader /></template>);
+      const { queryByRole } = await render(<template><SessionListHeader /></template>);
 
       // then
       assert.dom(queryByRole('link', { name: 'Créer/éditer plusieurs sessions' })).doesNotExist();
@@ -68,7 +68,7 @@ module('Integration | Component | panel-header', function (hooks) {
       this.owner.register('service:current-user', CurrentUserStub);
 
       // when
-      const screen = await render(<template><PanelHeader /></template>);
+      const screen = await render(<template><SessionListHeader /></template>);
 
       // then
       const createOneSessionButton = screen.getByRole('link', { name: 'Créer une session' });
